major refactor somewhat done, missing areas of effect again
