--- generate.bat
@@ -1,6 +1,0 @@
-node cleanUnityMetadata.js
-node getBreeds.js
-node getNameAndDescription.js
-node getEffectList.js
-node getSpellLevels.js
-node getTranslations.js
