--- generateFile.js
@@ -1,11 +1,0 @@
-const { getBreeds } = require('./getBreeds');
-const { getNameAndDescription } = require('./getNameAndDescription');
-const { getEffectList } = require('./getEffectList');
-const { getSpellLevels } = require('./getSpellLevels');
-const { getTranslations } = require('./getTranslations');
-
-getBreeds();
-getNameAndDescription();
-getEffectList();
-getSpellLevels();
-getTranslations();
