+++ readme.md
@@ -4,17 +4,23 @@ Script to generate a semi-automated version of the current spells in Dofus to si
 
 ## How to use
 
-First we must understand folder structure - we've got an input folder, a temp one and an output.
+First we must understand folder structure - we need an input folder and an output one.
 
-- Input will take a json version of the game files, which can be obtained with PyDofus.
-- Temp will generate some files needed for the script to run, for example reordering certain "breeds", as the game files reference them, and some spell pairs.
+- Input will take a json version of the game files, which can be obtained with [UABEA](https://github.com/nesrak1/UABEA) for now, until a better way has been found.
+- The script will generate some files it needs to run under the `./temp` folder and clean up the folder afterwards.
 - Output will be the autogenerated files that we can now use to edit DofusLab spell file (keeping in mind this is by no means a copy-pasteable version, there's some manual work involved like some duplicated effects on skills that have alternative effects, for example Morph, which will look like having 4 lines of damage).
 
 ### So, how do we use it then?
 
 - We must place a json copy of the following files into the `input` folder:
-`Effects.json, i18n_de.json, i18n_en.json, i18n_es.json, i18n_fr.json, i18n_it.json, i18n_pt.json, SpellLevels.json, SpellPairs.json, Spells.json, SpellStates.json, SpellVariants.json`.
+  `EffectsRoot.json, de.i18n.json, en.i18n.json, es.i18n.json, fr.i18n.json, it.i18n.json, pt.i18n.json, SpellLevelsRoot.json, SpellPairsRoot.json, SpellsRoot.json, SpellStatesRoot.json, SpellVariantsRoot.json`, `MonstersRoot.json`.
 
-- Execute `generate.bat` in order to generate the new data from the files.
+- Execute `node .\generate.js` in order to generate the new data from the files.
 
 - Then, the `output` folder will contain the `translatedEffects.json` file that we can use to start correlating with DofusLab files.
+
+### Current todos
+
+- Accept params to keep the temp folder for debug purposes (it does what it does in small steps so it's easy to pinpoint the location of things that fail, or to compare a step with the previous or the next).
+- Integrate unpacking the client with the script directly since UABEA is very manual. This could be done by downloading files from another unpacker in the community or handing it over to the application regardless.
+- Figure out areas of effect.
