--- temp/breedList.json
@@ -1,1 +1,0 @@
-[{"breed":1,"spells":[12977,13008,12978,13007,12981,13009,12976,13010,12994,13011,12979,13016,12992,13013,12993,13014,12984,13015,12995,13012,12983,13017,12986,13018,12988,13019,12980,13020,12987,13021,12989,13022,12990,13023,12997,13027,12991,13024,12985,13025,14434,14436,12982,13026]},{"breed":2,"spells":[13998,12618,12548,14021,12615,14017,12549,12600,12555,12604,12606,14008,12557,12608,12601,14018,14022,14023,12560,12609,14006,12559,12563,12614,14028,12619,12554,14009,12551,12612,14019,14020,12550,14015,12605,14016,18535,18656,14026,14025,14007,12617,12564,14010]},{"breed":3,"spells":[13338,13363,13336,13367,13328,13354,13331,14273,14278,13360,13342,13356,13334,13357,13340,13352,13332,13359,13364,13372,13344,13361,13339,13362,13365,13358,13337,13329,13335,13355,13366,13353,13343,13330,13349,14274,13346,13369,13333,13368,13345,13370,13347,13371]},{"breed":4,"spells":[12913,12930,12906,12935,12902,12932,12907,12933,12914,12922,12904,12939,12915,12936,12916,12937,12920,12938,12919,12942,12948,12943,12910,14314,12909,12940,12918,12911,12908,12945,12903,14741,12917,14742,14312,14313,12941,12905,12934,12950,12921,12949,12931,12947]},{"breed":5,"spells":[13249,13282,13244,13283,13242,13281,13254,13284,13243,13285,13253,13286,13287,13258,13245,13288,13261,13289,13259,13290,13246,13291,13257,13292,13251,13293,13248,13294,13250,13295,13247,14650,13255,13297,13256,13298,13252,13299,13260,14651,13262,13301,13296,13300]},{"breed":6,"spells":[12842,12863,12848,12864,12841,12866,14310,14311,12844,12865,12840,12871,12856,12869,12852,12867,12858,12870,12846,12872,12850,12868,12855,12873,12845,12874,12847,12875,12843,12876,12854,12877,12857,12878,12861,12883,12851,12880,12853,12881,12849,12879,12859,12882]},{"breed":7,"spells":[25795,25802,25858,25857,25877,25878,25870,25875,25742,25746,25867,25866,25868,25882,25871,25856,25860,25865,25744,25745,25879,25876,25880,25881,25743,25747,25863,25864,25873,25874,25749,25748,25859,25862,13184,13217,25869,25861,25872,25883,13175,25832,13187,13219]},{"breed":8,"spells":[13126,13135,13108,13148,13107,13143,14676,13130,13110,13115,13119,13138,13136,13142,13116,13137,13123,13117,13124,13156,13113,13140,13109,13141,13118,14677,14818,13133,13121,13144,13122,13145,13120,13112,13125,13146,13111,13131,13106,13139,13134,13147,13114,13132]},{"breed":9,"spells":[13055,13069,13047,13068,13050,13070,13051,13072,13058,13073,13053,13074,13054,13075,13066,13076,13056,13077,13057,13078,13052,13079,13059,13080,13062,13082,13061,13083,13064,13084,13049,13071,13063,13085,13065,13086,13067,13088,23510,13087,23499,23508,13060,14636]},{"breed":10,"spells":[13564,13515,13519,13560,13516,13559,14393,14396,13561,13526,13528,13563,13533,13562,13525,13565,13530,13566,13567,13522,13514,13568,13532,13569,13524,13571,13531,13570,13529,13572,13573,13523,13527,13574,13534,13579,13517,13576,13521,13577,13518,13575,13578,13520]},{"breed":11,"spells":[12737,12760,12744,12765,12746,12724,12733,12749,12736,12763,12745,12729,12756,12726,12761,14011,12738,12754,12734,12750,12731,12751,12739,12764,12723,12752,12728,12757,12759,12762,12743,14000,12727,12755,12732,12725,12730,12747,12735,12758,12741,12740,12748,12753]},{"breed":12,"spells":[12781,12815,12803,12805,12804,12780,12783,12806,12782,12786,12787,12810,12794,14309,12790,12814,12826,12777,12789,12813,12799,12824,12791,12807,12808,12821,12792,12816,12819,12785,12788,12820,12811,14307,12793,12809,12784,12823,24038,12817,12796,12797,12798,24039]},{"breed":13,"spells":[13432,13470,13444,13471,13435,13474,13434,13472,13433,13475,13437,13480,13440,13469,13439,13487,13436,13478,13429,13490,14415,13483,13477,13488,13431,13476,13442,14414,13445,14412,13438,13449,13441,13479,13450,13489,13446,13484,13473,14411,13443,13482,13491,13486]},{"breed":14,"spells":[13396,14681,13420,13408,13387,13406,13425,13405,13403,13410,13388,13414,13394,13419,13395,14682,13398,13411,13400,13412,13393,13390,13397,13417,13402,13385,13401,13418,13391,13389,13422,13407,13399,13426,13415,13421,13423,13392,18651,18650,13386,13409,13424,13404]},{"breed":15,"spells":[13823,13858,13827,13857,13829,13859,13832,13860,13822,13861,13834,13862,13819,13863,13830,13864,13821,13824,13820,13866,13826,13867,13838,13868,13831,13869,13837,13870,13836,13871,13865,13872,13835,13873,13849,13877,13825,13874,13828,13875,13833,14306,14308,13876]},{"breed":16,"spells":[14574,14604,14575,14602,14593,14603,14583,14605,14579,14606,14582,14607,14588,14618,14580,14609,14581,14590,14577,14611,14585,14612,14592,14613,14587,14614,14576,14637,14589,14616,14615,14617,14584,14608,14591,14622,14610,14620,14594,14619,14638,14595,14596,14621]},{"breed":17,"spells":[13668,13711,13669,13709,13670,13710,13666,13708,13667,13712,13674,13725,13679,13715,13696,13716,13680,13718,13676,13717,13695,13719,13682,13720,13681,13721,13701,13727,13677,13723,13684,13724,13678,13714,13675,13726,13697,13722,13683,13728,13672,13713,14341,14342]},{"breed":18,"spells":[13770,13806,13748,13786,13756,13787,13760,13788,13768,13796,13754,13795,13769,13791,13774,13792,13757,13789,13763,13794,13755,13804,13765,13753,13771,13797,13772,13798,13766,13799,13762,13793,13781,13801,13802,13758,13773,13803,13790,13800,13749,13805,14357,14321]},{"breed":19,"spells":[]},{"breed":20,"spells":[23432,23720,23719,23392,23801,23386,23263,23743,23824,23314,23736,24219,23393,23755,23330,23828,23754,23391,23741,23731,23434,23742,23726,23753,23823,23829,23747,23723,23401,23838,23729,23396,23826,23319,23268,23750,23841,23842,23738,23735,23834,23846,23734,23756]}]
