--- temp/modifiableEffectList.json
@@ -1,1 +1,0 @@
-[{"81":"#1{~1~2 to }#2 (healing)"},{"91":"#1{~1~2 to }#2 (Water steal)"},{"92":"#1{~1~2 to }#2 (Earth steal)"},{"93":"#1{~1~2 to }#2 (Air steal)"},{"94":"#1{~1~2 to }#2 (Fire steal)"},{"95":"#1{~1~2 to }#2 (Neutral steal)"},{"96":"#1{~1~2 to }#2 (Water damage)"},{"97":"#1{~1~2 to }#2 (Earth damage)"},{"98":"#1{~1~2 to }#2 (Air damage)"},{"99":"#1{~1~2 to }#2 (Fire damage)"},{"100":"#1{~1~2 to }#2 (Neutral damage)"},{"108":"#1{~1~2 to }#2 (Fire healing)"}]
